Split SwimmingMetrics.jsx
